backend/app/requirements.txt
# --- Core Web Framework ---
fastapi==0.115.0
uvicorn==0.30.6
requests==2.32.3

# --- ML / NLP Model ---
transformers==4.44.2
torch==2.4.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.2

# --- Utilities ---
python-dotenv==1.0.1
pydantic==2.8.2
